Reorganised the code for generating indexes.
